requirements.txt
typer
OpenEXR
numpy
pillow
tqdm